feat(rewash): scope-aware Excel exports for operator/wash-in/washer
- Extracted rewash export to utils/rewashExport.js (exportRewashExcel
  takes optional washerId to scope to a single washer; without it,
  exports every rewash_requests row).
- /washingin/rewash-download (existing): now shows ALL data (was
  filtered by req.user.id, which only matched if that user happened
  to be the requester — now supervisory view).
- /operator/rewash-download (new): operator-only, ALL data. Tile
  added to operator dashboard's Production category.
- /washingdashboard/rewash-download (new): washer-only, filtered to
  rewash_requests.washer_id = current user. Button "My Rewash" on the
  washing dashboard header.
