+++ views/operatorDashboard.ejs
@@ -733,6 +733,10 @@
           <i class="bi bi-stopwatch"></i>
           <span class="action-card-title">Lot TAT</span>
         </a>
+        <a href="/operator/rewash-download" class="action-card">
+          <i class="bi bi-arrow-repeat"></i>
+          <span class="action-card-title">Rewash Excel</span>
+        </a>
         <a href="/operator/editcuttinglots" class="action-card">
           <i class="bi bi-scissors"></i>
           <span class="action-card-title">Edit Lots</span>
