+++ views/washingDashboard.ejs
@@ -545,6 +545,9 @@
       <a href="/my-lots" class="btn-payments" style="background:#6366f1;">
         <i class="bi bi-bag-check"></i> My Lots
       </a>
+      <a href="/washingdashboard/rewash-download" class="btn-payments" style="background:#10b981;">
+        <i class="bi bi-file-earmark-excel"></i> My Rewash
+      </a>
       <a href="/payments/my-payments" class="btn-payments">
         <i class="bi bi-wallet2"></i> Payments
       </a>
